feat: add client tunnel tab (wip)
- add client tunnel tab
- merge server and client tunnel tab into TunnelManager component
- add related icons and translations
- add mock type for client tunnels
