--- src/components/standalone/openvpn_tunnel/ClientTunnel.vue
@@ -1,7 +1,0 @@
-<script setup lang="ts">
-import { NeTitle } from '@nethserver/vue-tailwind-lib'
-</script>
-
-<template>
-  <NeTitle>Client tunnel</NeTitle>
-</template>
