+++ src/components/standalone/openvpn_tunnel/DownloadTunnelModal.vue
@@ -7,11 +7,11 @@ import {
 } from '@nethserver/vue-tailwind-lib'
 import { ref, toRefs } from 'vue'
 import { useI18n } from 'vue-i18n'
-import type { ServerTunnelType } from './ServerTunnel.vue'
+import type { ServerTunnelType, ClientTunnelType } from './TunnelManager.vue'
 
 const props = defineProps<{
   visible: boolean
-  itemToDownload: ServerTunnelType | null
+  itemToDownload: ServerTunnelType | ClientTunnelType | null
 }>()
 
 const { t } = useI18n()
