+++ src/lib/fontawesome.ts
@@ -69,6 +69,7 @@ import { faCircleChevronUp } from '@fortawesome/free-solid-svg-icons/faCircleChe
 import { faDownLeftAndUpRightToCenter } from '@fortawesome/free-solid-svg-icons/faDownLeftAndUpRightToCenter'
 import { faUpRightAndDownLeftFromCenter } from '@fortawesome/free-solid-svg-icons/faUpRightAndDownLeftFromCenter'
 import { faCircleArrowDown } from '@fortawesome/free-solid-svg-icons/faCircleArrowDown'
+import { faCircleArrowUp } from '@fortawesome/free-solid-svg-icons/faCircleArrowUp'
 
 export async function loadFontAwesome(app: any) {
   app.component('FontAwesomeIcon', FontAwesomeIcon)
@@ -139,4 +140,5 @@ export async function loadFontAwesome(app: any) {
   library.add(faDownLeftAndUpRightToCenter)
   library.add(faUpRightAndDownLeftFromCenter)
   library.add(faCircleArrowDown)
+  library.add(faCircleArrowUp)
 }
